web-app/requirements.txt
pandas==1.5.3
Pillow==9.4.0
requests==2.28.2
streamlit==1.17.0
